// AMM/Izumi/types.ts
export type Token = {
    chainId: number
    address: string
    name: string
    symbol: string
    decimals: number
    logoURI: string
}

export type Pool = {
    tokenA: Token
    tokenB: Token
    pair: string
    fees: number
    tickSpacing: number,
    liquidity: bigint,
    sqrtPriceX96: bigint,
    tick: number,
    Quoter: Contract
    Pool: Contract
}
// ...
export type AddOptions = {
    max?: boolean
    slipage?: number
    deadline?: number
    tokenId?: number
}
// ...
export type Chains = 'arbitrum' | 'polygon' | 'optimism' | 'ethereum' | 'bsc' | 'polygonTestnet' | 'arbitrumTestnet' |
                    'ethereumTestnet' | 'optimismTestnet'